Moving sequence drawing functions to a separate file
