+++ sequences.py
@@ -1,11 +1,11 @@
 import warnings
 import copy
 import numpy as np
-import matplotlib.pyplot as plt
 from collections import namedtuple
 
 from devices import PasqalDevice
-from pulses import Pulse, ConstantWaveform
+from pulses import Pulse
+from seq_drawer import draw_sequence
 from utils import validate_duration
 
 # Auxiliary class to store the information in the schedule
@@ -73,10 +73,9 @@ class Sequence:
         self._taken_channels.append(channel_id)
         self._schedule[name] = []
 
-        if ch.basis_states not in self._phase_ref:
-            self._phase_ref[ch.basis_states] = {q: PhaseTracker(0)
-                                                for q in self._qids}
-            self._last_used[ch.basis_states] = {q: 0 for q in self._qids}
+        if ch.basis not in self._phase_ref:
+            self._phase_ref[ch.basis] = {q: PhaseTracker(0) for q in self._qids}
+            self._last_used[ch.basis] = {q: 0 for q in self._qids}
 
         if ch.addressing == 'global':
             self._schedule[name].append(TimeSlot('target', -1, 0, self._qids))
@@ -112,7 +111,7 @@ class Sequence:
                              "protocols: " + ", ".join(valid_protocols))
 
         t0 = last.tf    # Preliminary ti
-        basis = self._channels[channel].basis_states
+        basis = self._channels[channel].basis
         phase_barriers = [self._phase_ref[basis][q].last_time
                           for q in last.targets]
         current_max_t = max(t0, *phase_barriers)
@@ -175,7 +174,7 @@ class Sequence:
         elif len(qs) != 1:
             raise ValueError("This channel takes only a single target qubit.")
 
-        basis = self._channels[channel].basis_states
+        basis = self._channels[channel].basis
         phase_refs = {self._phase_ref[basis][q].last_phase for q in qs}
         if len(phase_refs) != 1:
             raise ValueError("Cannot target multiple qubits with different "
@@ -212,10 +211,10 @@ class Sequence:
 
         Args:
             basis (str): Valid basis for measurement (consult the
-                'supported_basis_states' attribute of the selected device for
+                'supported_bases' attribute of the selected device for
                 the available options).
         """
-        available = self._device.supported_basis_states
+        available = self._device.supported_bases
         if basis not in available:
             raise ValueError(f"The basis '{basis}' is not support by the "
                              "selected device. The available options are: "
@@ -258,149 +257,7 @@ class Sequence:
             self._phase_ref[basis][q][t] = new_phase
 
     def draw(self):
-        """Draw the entire sequence."""
-
-        def phase_str(phi):
-            value = (((phi + np.pi) % (2*np.pi)) - np.pi) / np.pi
-            if value == -1:
-                return r"$\pi$"
-            elif value == 0:
-                return "0"
-            else:
-                return r"{:.2g}$\pi$".format(value)
-
-        n_channels = len(self._channels)
-        if not n_channels:
-            raise SystemError("Can't draw an empty sequence.")
-        data = self._gather_data()
-        time_scale = 1e3 if self._total_duration > 1e4 else 1
-
-        # Boxes for qubit and phase text
-        q_box = dict(boxstyle="round", facecolor='orange')
-        ph_box = dict(boxstyle="round", facecolor='ghostwhite')
-
-        fig = plt.figure(constrained_layout=False, figsize=(20, 4.5*n_channels))
-        gs = fig.add_gridspec(n_channels, 1, hspace=0.075)
-
-        ch_axes = {}
-        for i, (ch, gs_) in enumerate(zip(self._channels, gs)):
-            ax = fig.add_subplot(gs_)
-            ax.spines['top'].set_color('none')
-            ax.spines['bottom'].set_color('none')
-            ax.spines['left'].set_color('none')
-            ax.spines['right'].set_color('none')
-            ax.tick_params(labelcolor='w', top=False, bottom=False, left=False,
-                           right=False)
-            ax.set_ylabel(ch, labelpad=40, fontsize=18)
-            subgs = gs_.subgridspec(2, 1, hspace=0.)
-            ax1 = fig.add_subplot(subgs[0, :])
-            ax2 = fig.add_subplot(subgs[1, :])
-            ch_axes[ch] = (ax1, ax2)
-            for j, ax in enumerate(ch_axes[ch]):
-                ax.axvline(0, linestyle='--', linewidth=0.5, color='grey')
-                if j == 0:
-                    ax.spines['bottom'].set_visible(False)
-                else:
-                    ax.spines['top'].set_visible(False)
-
-                if i < n_channels - 1 or j == 0:
-                    ax.tick_params(axis='x', which='both', bottom=True,
-                                   top=False, labelbottom=False, direction='in')
-                else:
-                    unit = 'ns' if time_scale == 1 else r'$\mu s$'
-                    ax.set_xlabel(f't ({unit})', fontsize=12)
-
-        for ch, (a, b) in ch_axes.items():
-            basis = self._channels[ch].basis_states
-            t = np.array(data[ch]['time']) / time_scale
-            ya = data[ch]['amp']
-            yb = data[ch]['detuning']
-
-            t_min = -t[-1]*0.03
-            t_max = t[-1]*1.05
-            a.set_xlim(t_min, t_max)
-            b.set_xlim(t_min, t_max)
-
-            max_amp = np.max(ya)
-            max_amp = 1 if max_amp == 0 else max_amp
-            amp_top = max_amp * 1.2
-            a.set_ylim(-0.02, amp_top)
-            det_max = np.max(yb)
-            det_min = np.min(yb)
-            det_range = det_max - det_min
-            if det_range == 0:
-                det_min, det_max, det_range = -1, 1, 2
-            det_top = det_max + det_range * 0.15
-            det_bottom = det_min - det_range * 0.05
-            b.set_ylim(det_bottom, det_top)
-
-            a.plot(t, ya, color="darkgreen", linewidth=0.8)
-            b.plot(t, yb, color='indigo', linewidth=0.8)
-            a.fill_between(t, 0, ya, color="darkgreen", alpha=0.3)
-            b.fill_between(t, 0, yb, color="indigo", alpha=0.3)
-            a.set_ylabel('Amplitude (MHz)', fontsize=12, labelpad=10)
-            b.set_ylabel('Detuning (MHz)', fontsize=12)
-
-            target_regions = []     # [[start1, [targets1], end1],...]
-            for coords in data[ch]['target']:
-                targets = list(data[ch]['target'][coords])
-                tgt_txt_y = max_amp*1.1-0.25*(len(targets)-1)
-                tgt_str = "\n".join([str(q) for q in targets])
-                if coords == 'initial':
-                    x = t_min + t[-1]*0.005
-                    target_regions.append([0, targets])
-                    if self._channels[ch].addressing == 'global':
-                        a.text(x, amp_top*0.98, "GLOBAL", fontsize=13,
-                               rotation=90, ha='left', va='top', bbox=q_box)
-                    else:
-                        a.text(x, tgt_txt_y, tgt_str, fontsize=12, ha='left',
-                               bbox=q_box)
-                        phase = self._phase_ref[basis][targets[0]][0]
-                        if phase:
-                            msg = r"$\phi=$" + phase_str(phase)
-                            a.text(0, max_amp*1.1, msg, ha='left', fontsize=12,
-                                   bbox=ph_box)
-                else:
-                    ti, tf = np.array(coords) / time_scale
-                    target_regions[-1].append(ti)   # Closing previous regions
-                    target_regions.append([tf, targets])  # Starting a new one
-                    phase = self._phase_ref[basis][targets[0]][tf * time_scale]
-                    a.axvspan(ti, tf, alpha=0.4, color='grey', hatch='//')
-                    b.axvspan(ti, tf, alpha=0.4, color='grey', hatch='//')
-                    a.text(tf + t[-1]*0.008, tgt_txt_y, tgt_str, ha='right',
-                           fontsize=12, bbox=q_box)
-                    if phase:
-                        msg = r"$\phi=$" + phase_str(phase)
-                        a.text(tf + t[-1]*0.02, max_amp*1.1, msg, ha='left',
-                               fontsize=12, bbox=ph_box)
-            # Terminate the last open region
-            target_regions[-1].append(t[-1])
-            for start, targets, end in target_regions:
-                q = targets[0]  # All targets have the same ref, so we pick
-                ref = self._phase_ref[basis][q]
-                for t_, delta in ref.changes(start, end, time_scale=time_scale):
-                    conf = dict(linestyle='--', linewidth=1.5, color='black')
-                    a.axvline(t_, **conf)
-                    b.axvline(t_, **conf)
-                    msg = u"\u27F2 " + phase_str(delta)
-                    a.text(t_, max_amp*1.1, msg, ha='right', fontsize=14,
-                           bbox=ph_box)
-
-            if 'measurement' in data[ch]:
-                msg = f"Basis: {data[ch]['measurement']}"
-                b.text(t[-1]*1.025, det_top, msg, ha='center', va='center',
-                       fontsize=14, color='white', rotation=90)
-                a.axvspan(t[-1], t_max, color='midnightblue', alpha=1)
-                b.axvspan(t[-1], t_max, color='midnightblue', alpha=1)
-                a.axhline(0, xmax=0.95, linestyle='-', linewidth=0.5,
-                          color='grey')
-                b.axhline(0, xmax=0.95, linestyle=':', linewidth=0.5,
-                          color='grey')
-            else:
-                a.axhline(0, linestyle='-', linewidth=0.5, color='grey')
-                b.axhline(0, linestyle=':', linewidth=0.5, color='grey')
-
-        plt.show()
+        draw_sequence(self)
 
     def __str__(self):
         full = ""
@@ -409,7 +266,7 @@ class Sequence:
         delay_line = "t: {}->{} | Delay \n"
         # phase_line = "t: {} | Phase shift of: {:.3f} | Targets: {}\n"
         for ch, seq in self._schedule.items():
-            basis = self._channels[ch].basis_states
+            basis = self._channels[ch].basis
             full += f"Channel: {ch}\n"
             first_slot = True
             for ts in seq:
@@ -464,54 +321,6 @@ class Sequence:
         if np.any(np.abs(pulse.detuning.samples) > ch.max_abs_detuning):
             raise ValueError("The pulse's detuning values go out of the range "
                              "allowed for the chosen channel.")
-
-    def _gather_data(self):
-        """Collects the whole sequence data for plotting."""
-        # The minimum time axis length is 100 ns
-        self._total_duration = max([self._last(ch).tf for ch in self._schedule
-                                    if self._schedule[ch]] + [100])
-        data = {}
-        for ch, seq in self._schedule.items():
-            time = [-1]     # To not break the "time[-1]" later on
-            amp = []
-            detuning = []
-            target = {}
-            # phase_shift = {}
-            for slot in seq:
-                if slot.ti == -1:
-                    target['initial'] = slot.targets
-                    time += [0]
-                    amp += [0]
-                    detuning += [0]
-                    continue
-                if slot.type in ['delay', 'target']:
-                    time += [slot.ti, slot.tf-1]
-                    amp += [0, 0]
-                    detuning += [0, 0]
-                    if slot.type == 'target':
-                        target[(slot.ti, slot.tf-1)] = slot.targets
-                    continue
-                pulse = slot.type
-                if (isinstance(pulse.amplitude, ConstantWaveform) and
-                        isinstance(pulse.detuning, ConstantWaveform)):
-                    time += [slot.ti, slot.tf-1]
-                    amp += [pulse.amplitude._value] * 2
-                    detuning += [pulse.detuning._value] * 2
-                else:
-                    time += list(range(slot.ti, slot.tf))
-                    amp += pulse.amplitude.samples.tolist()
-                    detuning += pulse.detuning.samples.tolist()
-            if time[-1] < self._total_duration - 1:
-                time += [time[-1]+1, self._total_duration-1]
-                amp += [0, 0]
-                detuning += [0, 0]
-            # Store everything
-            time.pop(0)     # Removes the -1 in the beginning
-            data[ch] = {'time': time, 'amp': amp, 'detuning': detuning,
-                        'target': target}
-            if hasattr(self, "_measurement"):
-                data[ch]['measurement'] = self._measurement
-        return data
 
 
 class PhaseTracker:
